requirements.txt
langchain
langchain-openai
langchain-community
chromadb
tiktoken
sqlalchemy
pandas
python-dotenv
PyMuPDF
pillow
tenacity
streamlit
python-docx
watchdog
